+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Fixed
+- No longer throw `Failed to reindex: No parameter type named ***` for custom parameter types.
+- Let the user choose in what file to generate step definitions.
+
 ## [0.12.1] - 2022-05-24
 ### Fixed
 - Fixed a bug with snippet generation